add calculations controller

+++ app/models/calculation.rb
@@ -2,7 +2,7 @@ class Calculation
   include Mongoid::Document
   include Mongoid::Timestamps
   field :operation, type: String
-  field :result, type: Integer, default: 0
+  field :result, type: Float, default: 0
   field :count, type: Integer, default: 1
   index({ operation: 1 }, { unique: true })
 end
